requirements.txt
# Core web framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# Data handling
PyYAML==6.0.1

# HTTP clients
httpx==0.27.0
curl_cffi>=0.7.0
requests>=2.31.0

# HTML parsing
beautifulsoup4>=4.12.0

# Cryptography
coincurve>=18.0.0
bcrypt>=4.1.0

# Database
asyncpg>=0.29.0

# Resilience
tenacity>=8.2.0

# Observability
prometheus-client>=0.19.0
sentry-sdk[fastapi]>=1.40.0
